+++ sentry-samples/sentry-samples-spring/src/main/java/io/sentry/samples/spring/AppConfig.java
@@ -1,0 +1,8 @@
+package io.sentry.samples.spring;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+
+@Configuration
+@Import(SentryConfig.class)
+public class AppConfig {}