+++ sentry-spring/src/main/java/io/sentry/spring/SentryRequestResolver.java
@@ -1,7 +1,7 @@
 package io.sentry.spring;
 
 import com.jakewharton.nopen.annotation.Open;
-import io.sentry.SentryOptions;
+import io.sentry.IHub;
 import io.sentry.protocol.Request;
 import io.sentry.util.Objects;
 import java.util.Arrays;
@@ -20,10 +20,10 @@ public class SentryRequestResolver {
   private static final List<String> SENSITIVE_HEADERS =
       Arrays.asList("X-FORWARDED-FOR", "AUTHORIZATION", "COOKIE");
 
-  private final @NotNull SentryOptions options;
+  private final @NotNull IHub hub;
 
-  public SentryRequestResolver(@NotNull SentryOptions options) {
-    this.options = Objects.requireNonNull(options, "options is required");
+  public SentryRequestResolver(final @NotNull IHub hub) {
+    this.hub = Objects.requireNonNull(hub, "options is required");
   }
 
   // httpRequest.getRequestURL() returns StringBuffer which is considered an obsolete class.
@@ -35,7 +35,7 @@ public class SentryRequestResolver {
     sentryRequest.setUrl(httpRequest.getRequestURL().toString());
     sentryRequest.setHeaders(resolveHeadersMap(httpRequest));
 
-    if (options.isSendDefaultPii()) {
+    if (hub.getOptions().isSendDefaultPii()) {
       sentryRequest.setCookies(toString(httpRequest.getHeaders("Cookie")));
     }
     return sentryRequest;
@@ -46,7 +46,7 @@ public class SentryRequestResolver {
     final Map<String, String> headersMap = new HashMap<>();
     for (String headerName : Collections.list(request.getHeaderNames())) {
       // do not copy personal information identifiable headers
-      if (options.isSendDefaultPii()
+      if (hub.getOptions().isSendDefaultPii()
           || !SENSITIVE_HEADERS.contains(headerName.toUpperCase(Locale.ROOT))) {
         headersMap.put(headerName, toString(request.getHeaders(headerName)));
       }