+++ sentry-spring/src/test/kotlin/io/sentry/spring/SentrySpringIntegrationTest.kt
@@ -154,6 +154,9 @@ open class App {
 
     @Bean
     open fun mockTransport() = transport
+
+    @Bean
+    open fun sentrySpringRequestListener() = SentrySpringRequestListener()
 }
 
 @RestController