+++ sentry/src/test/java/io/sentry/SentryClientTest.kt
@@ -69,7 +69,7 @@ class SentryClientTest {
             whenever(factory.create(any(), any())).thenReturn(transport)
         }
 
-        var attachment = Attachment("hello".toByteArray(), "hello.txt")
+        var attachment = Attachment("hello".toByteArray(), "hello.txt", "text/plain", true)
 
         fun getSut() = SentryClient(sentryOptions)
     }
@@ -764,6 +764,17 @@ class SentryClientTest {
     }
 
     @Test
+    fun `when captureTransaction with attachments not added to transaction`() {
+        val transaction = SentryTransaction("a-transaction")
+
+        val scope = createScopeWithAttachments()
+        scope.addAttachment(Attachment("hello".toByteArray(), "application/octet-stream"))
+        fixture.getSut().captureTransaction(transaction, scope, null)
+
+        verifyAttachmentsInEnvelope(transaction.eventId)
+    }
+
+    @Test
     fun `when scope's active span is a transaction, transaction context is applied to an event`() {
         val event = SentryEvent()
         val sut = fixture.getSut()
@@ -813,6 +824,27 @@ class SentryClientTest {
         assertEquals("transactionEnvironment", transaction.environment)
     }
 
+    @Test
+    fun `when transaction does not have tags, and tags are set on options, options values are applied to transactions`() {
+        fixture.sentryOptions.setTag("tag1", "value1")
+        val sut = fixture.getSut()
+        val transaction = SentryTransaction("name")
+        sut.captureTransaction(transaction)
+        assertEquals(mapOf("tag1" to "value1"), transaction.tags)
+    }
+
+    @Test
+    fun `when transaction has tags, and tags are set on options, options tags are added to transactions`() {
+        fixture.sentryOptions.setTag("tag1", "value1")
+        fixture.sentryOptions.setTag("tag2", "value2")
+        val sut = fixture.getSut()
+        val transaction = SentryTransaction("name")
+        transaction.setTag("tag3", "value3")
+        transaction.setTag("tag2", "transaction-tag")
+        sut.captureTransaction(transaction)
+        assertEquals(mapOf("tag1" to "value1", "tag2" to "transaction-tag", "tag3" to "value3"), transaction.tags)
+    }
+
     private fun createScope(): Scope {
         return Scope(SentryOptions()).apply {
             addBreadcrumb(Breadcrumb().apply {
@@ -834,7 +866,7 @@ class SentryClientTest {
             addAttachment(fixture.attachment)
 
             val bytesTooBig = ByteArray((fixture.maxAttachmentSize + 1).toInt()) { 0 }
-            addAttachment(Attachment(bytesTooBig, "will_get_dropped.txt"))
+            addAttachment(Attachment(bytesTooBig, "will_get_dropped.txt", "application/octet-stream", true))
         }
     }